+++ test/realm_test.dart
@@ -1345,11 +1345,14 @@ Future<void> main([List<String>? args]) async {
     final configuration = Configuration.flexibleSync(user, [Task.schema]);
 
     int transferredBytes = -1;
-
+    final completer = Completer<void>();
     var syncedRealm = await getRealmAsync(configuration, onProgressCallback: (syncProgress) {
       transferredBytes = syncProgress.transferredBytes;
+      if (syncProgress.transferredBytes == syncProgress.transferableBytes) {
+        completer.complete();
+      }
     });
-
+    completer.future.timeout(Duration(milliseconds: 300), onTimeout: () => throw Exception("onProgressCallback did not happen."));
     expect(syncedRealm.isClosed, false);
     expect(transferredBytes, greaterThan(-1));
   });